Bug 1822378 - Refactor OnboardingInteractor into its own file
