+++ fenix/app/src/main/java/org/mozilla/fenix/home/sessioncontrol/SessionControlInteractor.kt
@@ -27,6 +27,7 @@ import org.mozilla.fenix.home.recentvisits.RecentlyVisitedItem.RecentHistoryGrou
 import org.mozilla.fenix.home.recentvisits.RecentlyVisitedItem.RecentHistoryHighlight
 import org.mozilla.fenix.home.recentvisits.controller.RecentVisitsController
 import org.mozilla.fenix.home.recentvisits.interactor.RecentVisitsInteractor
+import org.mozilla.fenix.onboarding.interactor.OnboardingInteractor
 import org.mozilla.fenix.search.toolbar.SearchSelectorInteractor
 import org.mozilla.fenix.search.toolbar.SearchSelectorMenu
 import org.mozilla.fenix.wallpapers.WallpaperState
@@ -147,30 +148,6 @@ interface ToolbarInteractor {
     fun onPaste(clipboardText: String)
 }
 
-/**
- * Interface for onboarding related actions in the [SessionControlInteractor].
- */
-interface OnboardingInteractor {
-    /**
-     * Hides the onboarding and navigates to Search. Called when a user clicks on the "Start Browsing" button.
-     */
-    fun onStartBrowsingClicked()
-
-    /**
-     * Opens a custom tab to privacy notice url. Called when a user clicks on the "read our privacy notice" button.
-     */
-    fun onReadPrivacyNoticeClicked()
-
-    /**
-     * Show Wallpapers onboarding dialog to onboard users about the feature if conditions are met.
-     * Returns true if the call has been passed down to the controller.
-     *
-     * @param state The wallpaper state.
-     * @return Whether the onboarding dialog is currently shown.
-     */
-    fun showWallpapersOnboardingDialog(state: WallpaperState): Boolean
-}
-
 interface CustomizeHomeIteractor {
     /**
      * Opens the customize home settings page.
@@ -238,6 +215,20 @@ interface MessageCardInteractor {
 }
 
 /**
+ * Interface for wallpaper related actions.
+ */
+interface WallpaperInteractor {
+    /**
+     * Show Wallpapers onboarding dialog to onboard users about the feature if conditions are met.
+     * Returns true if the call has been passed down to the controller.
+     *
+     * @param state The wallpaper state.
+     * @return Whether the onboarding dialog is currently shown.
+     */
+    fun showWallpapersOnboardingDialog(state: WallpaperState): Boolean
+}
+
+/**
  * Interactor for the Home screen. Provides implementations for the CollectionInteractor,
  * OnboardingInteractor, TopSiteInteractor, TabSessionInteractor, ToolbarInteractor,
  * ExperimentCardInteractor, RecentTabInteractor, RecentBookmarksInteractor
@@ -263,7 +254,8 @@ class SessionControlInteractor(
     RecentVisitsInteractor,
     CustomizeHomeIteractor,
     PocketStoriesInteractor,
-    SearchSelectorInteractor {
+    SearchSelectorInteractor,
+    WallpaperInteractor {
 
     override fun onCollectionAddTabTapped(collection: TabCollection) {
         controller.handleCollectionAddTabTapped(collection)
