+++ fenix/app/src/main/java/org/mozilla/fenix/onboarding/interactor/OnboardingInteractor.kt
@@ -1,0 +1,20 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+package org.mozilla.fenix.onboarding.interactor
+
+/**
+ * Interface for onboarding related actions.
+ */
+interface OnboardingInteractor {
+    /**
+     * Hides the onboarding and navigates to Search. Called when a user clicks on the "Start Browsing" button.
+     */
+    fun onStartBrowsingClicked()
+
+    /**
+     * Opens a custom tab to privacy notice url. Called when a user clicks on the "read our privacy notice" button.
+     */
+    fun onReadPrivacyNoticeClicked()
+}
